refactor: move remaining of output.text.message to core.text_utils
